leetcode-qns-service/requirements.txt
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
google-api-core==2.12.0
google-auth==2.23.4
google-cloud==0.34.0
google-cloud-appengine-logging==1.3.2
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.3
google-cloud-logging==3.8.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.2
grpcio-status==1.59.2
idna==3.4
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
urllib3==2.0.7

